Fix YAML spec compliance gaps across models, controllers, and Lua scripts
Key fixes:
- Subject: Remove @NotBlank on tenant (YAML uses anyOf, not required tenant);
  add validateSubject() to enforce at least one standard field
- Enums: Rename ReservationState -> ReservationStatus to match YAML schema name;
  add DecisionEnum (ALLOW, ALLOW_WITH_CAPS, DENY)
- DecisionResponse/ReservationCreateResponse: Use DecisionEnum instead of String
- Caps: Change Long -> Integer for max_tokens/max_steps_remaining/cooldown_ms
  (YAML type: integer without int64 format); add @Size(max=256) on tool lists
- StandardMetrics: Change Long -> Integer for tokens_input/tokens_output/latency_ms
- SignedAmount: Add @NotNull on required unit/amount fields
- EventCreateRequest: Add @Min(0) on client_time_ms per YAML minimum constraint
- ReservationExtendRequest: Use @Min(1) instead of @Positive for clarity
- event.lua: Add UNIT_MISMATCH validation (spec requires 400 for unit mismatch)
- Repository: Update all string decision literals to use DecisionEnum constants;
  update ReservationState references to ReservationStatus

+++ cycles-protocol-service/cycles-protocol-service-model/src/main/java/io/runcycles/protocol/model/StandardMetrics.java
@@ -8,9 +8,9 @@ import java.util.Map;
 /** Cycles Protocol v0.1.23 */
 @Data @Builder @NoArgsConstructor @AllArgsConstructor @JsonInclude(JsonInclude.Include.NON_NULL)
 public class StandardMetrics {
-    @Min(0) @JsonProperty("tokens_input") private Long tokensInput;
-    @Min(0) @JsonProperty("tokens_output") private Long tokensOutput;
-    @Min(0) @JsonProperty("latency_ms") private Long latencyMs;
+    @Min(0) @JsonProperty("tokens_input") private Integer tokensInput;
+    @Min(0) @JsonProperty("tokens_output") private Integer tokensOutput;
+    @Min(0) @JsonProperty("latency_ms") private Integer latencyMs;
     @Size(max = 128) @JsonProperty("model_version") private String modelVersion;
     @JsonProperty("custom") private Map<String, Object> custom;
 }
